Add feature for search/replace of cropped image file names within post content, in addition to the existing search/replace of original file names.
Previously, doing a "Replace the file, use new file name and update all links" operation only replaced the original image filename in post content, meaning that when cropped images (with links to their full-size version) existed in posts, the link to the full image was updated but the image in the post still pointed to the old image which had been deleted, so the image was left broken.

Updating the source URL for the cropped image aligns better with a user's intent for replacing the image in the first place.