Check if replace_and_search is allowed
Uses an explicit check at the requested type of action rather than
handling the default via else. A filter is also added to this check
that allows replace_and_search to be disabled so that only replace
can be used.

This filter - emr_enable_replace_and_search - can be used elsewhere
to show or hide replace and search options.